nextUser.js fully refactored
